autotools: add `check-progs` target
If you configure with "--enable-tests=no", then the tests
are not build during `make all`, but only during `make check`.
That is convenient for development, so if you don't work on tests,
they usually don't get build and the (partial) compilation is faster.

However, there is no target to build the tests without running them.

Add `make check-progs` for that.