shared: copy file instead of symlink "nm-ethtool-utils.h"
"nm-ethtool-utils.h" is part of public API in libnm(-core). It is also
used by shared/nm-base, which must have no dependency on libnm-core.

This was previously solved by symlinking the file. I find that error
prone, because the user might edit one file, without realizing that the
other file also changes.

Instead, copy the file and have it twice. Note that we have a unit test
which checks that both files are (and state) the same.

Copy+paste is a valid way of reusing code. By checking that the copy
does not diverge from the original, the downsides of copy+paste are
mitigated.