build/meson: include subdir "libnm" before "src/core"
No big change, but eventually I' like to move all source
directories under src/. That must be done one after the other,
so the first step is to move libnm-core/ into src/. If libnm
gets loaded in between, that causes odd ordering.

"src/core" should not depend on "libnm" and vice versa, so this
should have little effect for now.