libnm: add API to NMLldpNeighbor to access by index
NMLldpNeighbor tries to abstract the way how tracking of the attributes
is implemented. For example, it does not offer an API to lookup by index
and it (previously) did not guarantee any sort order of the attribute
names.

Let's be more explicit about what and how the attributes are tracked. NMLldpNeighbor
is a list of name-gvariant pairs. Let's commit to some things:

 - the instance shall be immutable and there is no API for changing them
   (or even for creating them, aside nm_lldp_neighbor_new(), which is pointless).

 - when we expose the list of names, the names shall be sorted.

 - as we sort them, each attribute has a fixed index.

Add API to access the attributes by index. Also add API to access the
list of names without requiring to copy the list.

While at it, also rework the way we track the data. NMLldpNeighbor is
simple, the attributes don't change. We don't need a GHashTable for
that, instead track the attributes in a way that is more suited to the
way we use it. Since we have a sorted list of names, we still can do
fast binary search to find an attribute by name.