m4: ax_lib_readline.m4: Use readline libraries when linking add_history test program
Up till now the AC_LANG_CALL() was just checking if program which
uses the add_history call can be correctly linked without passing
required libraries (e.g. -libreadline to LIBS variable).

This commit fixes this issue as now the add_history can be found as it
is now available in any of passed libs.