initrd: avoid cloning string in reader_parse_rd_znet()
The code did:

   key    = g_strndup(tmp, val - tmp);
   val[0] = '\0';

That is pointless. If we strndup the key, we don't need to truncate
the string at the '='. It might be nicer not to mutate the input string,
however, the entire code with "argument" parsing is about mutating the
input string, so that is something we apparently are fine with.

As such, don't clone the string anymore.